+++ extension/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Unreleased
+
+- Renamed **Private Extensions: Clean NPM Cache** to **Private Extensions: Garbage Collect NPM Cache**.
+- Added a **Private Extensions: Delete NPM Cache** command to completely delete the cache.
+
 ## 1.3.0-beta.0
 
 - Upgraded [npm-registry-fetch](https://github.com/npm/npm-registry-fetch#-fetch-options) again